# laravel9-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto V_Lab",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transactions/{user_id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Listar transações de um usuário",
                "description": "Listar todas as transações feitas por um usuário.",
                "operationId": "e04cd602b5684128605ad80a1935f84b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de transações do usuário"
                    },
                    "404": {
                        "description": "Nenhuma transação encontrada para o usuário"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Criar transação",
                "description": "Criar uma nova transação para o usuário.",
                "operationId": "d32ce615b35948ce7d42dfa17db1f1bf",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Despesa"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transação criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação dos dados"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Listar todas as transações de todos os usuários",
                "description": "Listar todas as transações feitas por todos os usuários.",
                "operationId": "9b5aba419772e977aaba996b5ba2ad67",
                "responses": {
                    "200": {
                        "description": "Lista de todas as transações de todos os usuários"
                    },
                    "404": {
                        "description": "Nenhuma transação encontrada"
                    }
                }
            }
        },
        "/api/transactions/{user_id}/delete/{transaction_id}": {
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Apagar transação",
                "description": "Apagar uma transação do usuário.",
                "operationId": "3537b36b9e2c7e1e3064393276e5a323",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "description": "ID da transação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transação apagada com sucesso"
                    },
                    "404": {
                        "description": "Transação não encontrada"
                    }
                }
            }
        },
        "/api/users/{id}/delete": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Apagar usuário",
                "description": "Apagar um usuário pelo seu ID",
                "operationId": "80d42eb320ae7317752d2784434df23b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a ser apagado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}